This PR makes imdct faster and decoding a little faster.

Before this PR (on MacBook Pro 2015):

```
$ go test -bench=Bench
BenchmarkSetup-4            1000           1280923 ns/op
BenchmarkDecode-4           1000           2063860 ns/op
BenchmarkIMDCT-4           10000            112480 ns/op
PASS
ok      github.com/jfreymuth/vorbis     4.854s
```

After this PR:

```
$ go test -bench=Bench
BenchmarkSetup-4            1000           1270706 ns/op
BenchmarkDecode-4           1000           1886631 ns/op
BenchmarkIMDCT-4           20000             92588 ns/op
PASS
ok      github.com/jfreymuth/vorbis     6.298s
```